rename to display clusters
