+++ display-clusters/build.gradle.kts
@@ -7,7 +7,7 @@ android {
     compileSdk = libs.versions.compileSdk.get().toInt()
 
     defaultConfig {
-        applicationId = "com.esri.arcgismaps.sample.displaypointsusingclusteringfeaturereduction"
+        applicationId = "com.esri.arcgismaps.sample.displayclusters"
         minSdk = libs.versions.minSdk.get().toInt()
         targetSdk = libs.versions.targetSdk.get().toInt()
         versionCode = libs.versions.versionCode.get().toInt()
@@ -31,7 +31,7 @@ android {
         kotlinCompilerExtensionVersion = libs.versions.kotlinCompilerExt.get()
     }
 
-    namespace = "com.esri.arcgismaps.sample.displaypointsusingclusteringfeaturereduction"
+    namespace = "com.esri.arcgismaps.sample.displayclusters"
 }
 
 dependencies {
