+++ display-clusters/src/main/AndroidManifest.xml
