+++ display-clusters/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <resources>
-    <string name="app_name">Display points using clustering feature reduction</string>
+    <string name="app_name">Display clusters</string>
     <string name="portal_url">https://www.arcgis.com/</string>
 </resources>
